% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cnp_is_valid.R
\name{check_cnp_is_valid}
\alias{check_cnp_is_valid}
\title{Check if a CNP is valid}
\usage{
check_cnp_is_valid(cnp)
}
\arguments{
\item{cnp}{a 13-digit number (or a vector of 13-digit numbers) representing
the Personal Numeric Code - CNP}
}
\value{
a logical vector (\code{TRUE}, \code{FALSE} or \code{NA})
}
\description{
Check if a CNP is valid
}
\examples{
check_cnp_is_valid(1940616346114)
check_cnp_is_valid(7041218318525)
check_cnp_is_valid(62012060180783)
check_cnp_is_valid(NA)
}
