% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_lookup}
\alias{county_lookup}
\title{County lookup table}
\format{
\subsection{\code{county_lookup}}{

A data frame with 48 rows and 2 columns:
\describe{
\item{code}{county code}
\item{county}{county name}
}
}
}
\source{
\url{https://ro.wikipedia.org/wiki/Cod_numeric_personal_(România)#JJ}
}
\usage{
county_lookup
}
\description{
A lookup table mapping the county code to the county name.
}
\keyword{datasets}
