% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days_in_month}
\alias{days_in_month}
\title{Days in month lookup table}
\format{
\subsection{\code{days_in_month}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{month}{month as padded character - \code{"01"} to \code{"12"}}
\item{days_non_leap}{number of days in a non-leap year}
\item{days_leap}{number of days in a leap year}
}
}
}
\usage{
days_in_month
}
\description{
A lookup table containing the number of days in a month.
}
\keyword{datasets}
