% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sex}
\alias{extract_sex}
\alias{extract_birth_year}
\alias{extract_birth_month}
\alias{extract_dob}
\alias{extract_county}
\alias{extract_status}
\title{Extract CNP components}
\usage{
extract_sex(x)

extract_birth_year(x)

extract_birth_month(x)

extract_dob(x)

extract_county(x)

extract_status(x)
}
\arguments{
\item{x}{a \code{cnp} vector.}
}
\value{
\itemize{
\item \code{extract_sex()}: a character vector containing the parsed sex: \code{M} or \code{F}.
\item \code{extract_birth_year()}: a character vector containing the birth year.
\item \code{extract_birth_month()}: a character vector containing the birth month.
\item \code{extract_county()}: a character vector containing the county of issue.
\item \code{extract_status()}: a character vector containing the residence status
(\code{"native"} or \code{"resident"}).
\item \code{extract_dob()}: a vector of containing the date of birth.
}
}
\description{
The \code{extract_()} family of functions provides tools to extract the different
pieces of information encoded in the CNP. These functions return a character
vector:
\itemize{
\item \code{extract_sex()} extracts the sex (\code{"M"} or \code{"F"}) based on the \code{S}
component of the CNP.
\item \code{extract_birth_year()} extracts the year of birth, inferred from the \code{S}
(which also encodes the century of birth, alongside the sex ) and \code{AA}
(encoding the final 2 digits of the birth year) components of the CNP.
\item \code{extract_birth_month()} extracts the month of birth based on the \code{"LL"}
component.
\item \code{extract_county()} get the county of issue (based on the \code{JJ} component).
\item \code{extract_status()} extract the residence status (based on the \code{S}
component)
}

This function returns a date vector:
\itemize{
\item \code{extract_dob()} get the date of birth (based on the \code{S}, \code{AA}, \code{LL}, and
\code{ZZ} components).
}
}
\examples{

# Extract sex
extract_sex(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))

# Extract birth year
extract_birth_year(
  cnp(c("5201206346491", "1940616346114", "7041218318525", NA))
)

# Extract birth month
extract_birth_month(cnp(c(5201206346491, 1940616346114, 7041218318525, NA)))

# Extract date of birth
extract_dob(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))

# Extract county
extract_county(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))

# Extract residence status
extract_status(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))

}
