% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county.R
\name{get_county}
\alias{get_county}
\title{Extract County of Issue from the Personal Numeric Code}
\usage{
get_county(cnp)
}
\arguments{
\item{cnp}{(\code{cnp}) a \code{cnp} vector.}
}
\value{
a character vector containing the names of the counties where the
CNPs were issued.
}
\description{
Extract the county based on the \code{JJ} component of the CNP and parse it.
}
\examples{
get_county(cnp("1940616346114"))
get_county(cnp("7041218318525"))
get_county(cnp("6201206018078"))
get_county(cnp("5201206346491"))
get_county(cnp(c("5201206346491", "1940616346114", "7041218318525")))
get_county(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))
}
