% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dob.R
\name{get_dob}
\alias{get_dob}
\title{Extract the Date of Birth from the Personal Numeric Code}
\usage{
get_dob(cnp)
}
\arguments{
\item{cnp}{(\code{cnp}) a \code{cnp} vector.}
}
\value{
a vector of dates.
}
\description{
Extract the date of birth based on the relevant CNP components.
}
\examples{
get_dob(cnp("1940616346114"))
get_dob(cnp("7041218318525"))
get_dob(cnp("6201206018078"))
get_dob(cnp("5201206346491"))
get_dob(cnp(c("5201206346491", "1940616346114", "7041218318525")))
get_dob(cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))
}
