% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex.R
\name{get_sex}
\alias{get_sex}
\title{Extract the Sex from the Personal Numeric Code}
\usage{
get_sex(cnp)
}
\arguments{
\item{cnp}{a 13-digit number representing the Personal Numeric Code - CNP}
}
\value{
a character vector of \code{M} and/ or \code{F}
}
\description{
This function extracts the sex based on the sex component \code{S}. Worth noting
that, at the moment of the implementation, Romanian authorities view sex as
binary. In the event of a sex change a new CNP is issued.
}
\examples{
get_sex(1940616346114)
get_sex(7041218318525)
get_sex(6201206018078)
get_sex(5201206346491)
}
