% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Extract Residence Status from the Personal Numeric Code}
\usage{
get_status(cnp)
}
\arguments{
\item{cnp}{(\code{cnp}) a \code{cnp} vector.}
}
\value{
a character vector containing the residence status: \code{"native"} or
\code{"resident"}.
}
\description{
Extract Residence Status from the Personal Numeric Code
}
\examples{
get_status(cnp("6201206018078"))
get_status(cnp(c("5201206346491", "1940616346114")))
}
