% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_cnp.R
\name{interpret_cnp}
\alias{interpret_cnp}
\alias{interpreteaza_cnp}
\title{Interpret the Personal Numeric Code}
\usage{
interpret_cnp(cnp, lang = c("RO", "EN"))

interpreteaza_cnp(cnp, lang = c("RO", "EN"))
}
\arguments{
\item{cnp}{a 13-digit number (or a vector of 13-digit numbers) representing
the Personal Numeric Code - CNP}

\item{lang}{a string denoting the language for the interpretation - either
\code{RO} for Romanian (the default) or \code{EN} for English.}
}
\value{
a character interpretation the Personal Numeric Code
}
\description{
This function offers a verbose interpretation of the Personal Numeric Code.
}
\examples{
interpret_cnp(6201206018078)
interpret_cnp(5201206346491, lang = "RO")
interpret_cnp(1940616346114)
interpret_cnp(1940616346114, lang = "EN")
interpret_cnp(7041218318525)
}
