% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocnp-package.R
\docType{package}
\name{rocnp-package}
\alias{rocnp}
\alias{rocnp-package}
\title{rocnp: Tools for Working with Romanian Personal Numeric Codes PNC / CNP}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dragosmg/rocnp}
  \item Report bugs at \url{https://github.com/dragosmg/rocnp/issues}
}

}
\author{
\strong{Maintainer}: Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com}

}
\keyword{internal}
