% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_cnp.R
\name{augment_cnp}
\alias{augment_cnp}
\title{Parse the CNP fields}
\usage{
augment_cnp(x)
}
\arguments{
\item{x}{(list) a decomposed CNP.}
}
\value{
th input list with a series of additional fields:
\itemize{
\item \code{sex}: the parsed sex
\item \code{dob}: the reconstructed date of birth
}
}
\description{
Parse the CNP fields
}
\keyword{internal}
