% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp.R
\name{cnp}
\alias{cnp}
\title{Create a \code{cnp} object}
\usage{
cnp(x = character())
}
\arguments{
\item{x}{(character) a character (or values that can be cast to character)
vector of personal numeric codes (CNPs).}
}
\value{
a vector of class \code{cnp}.
}
\description{
Create a \code{cnp} object
}
\examples{
cnp(c("1940616346114", "7041218318525"))
}
