% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birth_month.R
\name{get_birth_month}
\alias{get_birth_month}
\title{Extract the Month of Birth from the Personal Numeric Code}
\usage{
get_birth_month(
  cnp,
  lang = c("RO", "EN"),
  type = c("long", "numeric", "short")
)
}
\arguments{
\item{cnp}{a 13-digit number representing the Personal Numeric Code - CNP}

\item{lang}{a string denoting the language for the interpretation - either
\code{RO} for Romanian (the default) or \code{EN} for English.}

\item{type}{string indicating type of output desired, one of \code{"long"}, full
month name (default), \code{"short"}, short month name, and \code{"numeric"}}
}
\value{
a string or numeric depending on the requested output
}
\description{
This function extracts the month of birth based on \code{"LL"}. The main purpose
of this function is to help the \code{interpret_cnp()} functionality. The year is
returned as string, in the specified language - defaults to Romanian.
}
\examples{
a <- decompose_cnp(1940616346114)
get_birth_month(a)
get_birth_month(a, type = "short")
get_birth_month(a, type = "numeric")
b <- decompose_cnp(7041218318525)
get_birth_month(b, lang = "EN")
get_birth_month(b, lang = "EN", type = "long")
get_birth_month(b, lang = "EN", type = "numeric")
get_birth_month(b, lang = "EN", type = "short")
}
