% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birth_year.R
\name{get_birth_year}
\alias{get_birth_year}
\title{Extract Year of Birth from the Personal Numeric Code}
\usage{
get_birth_year(cnp)
}
\arguments{
\item{cnp}{(\code{cnp}) a \code{cnp} vector.}
}
\value{
a character vector representing the year of birth.
}
\description{
Extract the year of birth, inferred starting from the \code{S} component, the
code for sex - which is also indicative of the century of birth - and the
\code{AA} component denoting the final 2 digits of the birth year. The year is
returned as string,
}
\examples{
get_birth_year(cnp("1940616346114"))
get_birth_year(cnp("7041218318525"))
}
