% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birth_year.R
\name{get_birth_year}
\alias{get_birth_year}
\title{Extract Birth Year from the Personal Numeric Code}
\usage{
get_birth_year(cnp_dec, lang = c("RO", "EN"))
}
\arguments{
\item{cnp_dec}{a named character vector representing a decomposed Personal
Numeric Code}

\item{lang}{a string denoting the language for the interpretation - either
\code{RO} for Romanian (the default) or \code{EN} for English.}
}
\value{
an string representing the year of birth or the possible years (in
the case of residents)
}
\description{
This function infers the year of birth based on the code for gender - which
is indicative of the century of birth - and the \code{AA} component denoting the
last 2 digits of the birth year. The main purpose of this function is to help
the \code{interpret_cnp()} functionality. Thus, the year is returned as string,
especially since for residents not born in Romania, there might be two
possible valid values for the birth year.
}
\examples{
a <- decompose_cnp(1940616346114)
get_birth_year(a)
b <- decompose_cnp(7041218318525)
get_birth_year(b)
}
