% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county.R
\name{get_county}
\alias{get_county}
\title{Extract County of Issue from the Personal Numeric Code}
\usage{
get_county(cnp)
}
\arguments{
\item{cnp}{a 13-digit number (or a vector of 13-digit numbers) representing
the Personal Numeric Code - CNP}
}
\value{
a string representing the name of the county where the CNP was issued
}
\description{
Extract County of Issue from the Personal Numeric Code
}
\examples{
get_county(6201206018078)
get_county(5201206346491)
get_county(1940616346114)
get_county(7041218318525)
}
