% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex.R
\name{get_sex}
\alias{get_sex}
\title{Extract Sex from the Personal Numeric Code}
\usage{
get_sex(cnp)
}
\arguments{
\item{cnp}{(\code{cnp}) a \code{cnp} vector.}
}
\value{
a character vector of the parsed sex: \code{M} or \code{F}.
}
\description{
Extract the sex based on the \code{S} component of the CNP.
}
\examples{
get_sex(new_cnp("1940616346114"))
get_sex(new_cnp("7041218318525"))
get_sex(new_cnp("6201206018078"))
get_sex(new_cnp("5201206346491"))
get_sex(new_cnp(c("5201206346491", "1940616346114", "7041218318525")))
get_sex(new_cnp(c("5201206346491", "1940616346114", "7041218318525", NA)))
}
