% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocnp-package.R
\docType{package}
\name{rocnp-package}
\alias{rocnp}
\alias{rocnp-package}
\title{rocnp: Work with Romanian Personal Numeric Codes PNC / CNP}
\description{
'rocnp' provides a set of tools for working with Romanian personal numeric codes. The central tool is a validation function which applies several verification criteria, as defined in the legislation, to assess the validity of the numeric code. This is accompanied by functionality for extracting the different components of a personal numeric code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dragosmg/rocnp}
  \item Report bugs at \url{https://github.com/dragosmg/rocnp/issues}
}

}
\author{
\strong{Maintainer}: Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com} [copyright holder]

}
\keyword{internal}
