% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocnp-package.R
\docType{package}
\name{rocnp-package}
\alias{rocnp}
\alias{rocnp-package}
\title{rocnp: Work with Romanian Personal Numeric Codes PNC / CNP}
\description{
A set of tools for working with Romanian personal numeric codes. The core is a validation function which applies several verification criteria to assess the validity of numeric codes. This is accompanied by functionality for extracting the different components of a personal numeric code. A personal numeric code is issued to all Romanian residents either at birth or when they obtain a residence permit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dragosmg/rocnp}
  \item Report bugs at \url{https://github.com/dragosmg/rocnp/issues}
}

}
\author{
\strong{Maintainer}: Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com} [copyright holder]

}
\keyword{internal}
